+++ src/lib/flows/pull-request-flow.js
@@ -18,17 +18,20 @@ async function start(context, isArchiveArtifacts = true) {
   core.startGroup(
     `[Pull Request Flow] Checking out ${context.config.github.groupProject} and its dependencies`
   );
+  const leafToGetTreeFrom = context.config.github.inputs.startingProject
+    ? context.config.github.inputs.startingProject
+    : context.config.github.repository;
   const definitionTree = await getTreeForProject(
     context.config.github.inputs.definitionFile,
-    context.config.github.repository,
+    leafToGetTreeFrom,
     await getPlaceHolders(context, context.config.github.inputs.definitionFile)
   );
   const nodeChain = await parentChainFromNode(definitionTree);
 
   logger.info(
-    `Tree for project ${
-      context.config.github.repository
-    }. Dependencies: ${nodeChain.map(node => "\n" + node.project)}`
+    `Tree for project ${leafToGetTreeFrom}. Dependencies: ${nodeChain.map(
+      node => "\n" + node.project
+    )}`
   );
   const checkoutInfo = await checkoutDefinitionTree(context, nodeChain);
   core.endGroup();