+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 
 ## Enhancements:
 
-- command line ready
 - token added to git commands to avoid private repositories problem from CLI execution
+- clone option added
 
 # V2.2